Ingo Karkat + Andy Wokula - BUG
Creation of literal pattern via '\V' in {Visual}<Leader>m mapping collided with individual escaping done in <Leader>m mapping so that an escaped '\*' would be interpreted as a multi item when both modes are used for marking. Replaced \V with s:EscapeText() to be consistent. Replaced the (overly) generic mark#GetVisualSelectionEscaped() with mark#GetVisualSelectionAsRegexp() and mark#GetVisualSelectionAsLiteralPattern(). Thanks to Andy Wokula for the patch. "