ENH: Use ingo#regexp#split#AddPatternByProjectedMatchLength() for a more correct insertion point with multiple alternatives for a mark
When adding a pattern (e.g. /foo/) to the same mark group when a shorter prefix (/fo/) already exists, the shorter one will eclipse the longer one (only "fo" will be matched in "foobar").
The new function determines / estimates (in case of multis) the match length, and adds shorter ones after longer ones, to avoid the problem. In case of varying match lengths (e.g. with * and \+ multis), it cannot completely fix it, but it's always equal or better than the original simple addition at the end.