BUG: Regression: <Leader>n without {N} and not on an existing mark prints error "Do not pass empty pattern to disable all marks
Instead of indiscriminately passing the current mark text to mark#DoMark(), only do so if a mark group is passed or we're actually on a mark. For disabling all marks, now call the one-argument version to avoid the error message (that was meant for the :Mark command, but also affected mark clearing as an unintended side effect).

Fixes #14.