Refactoring: Combine mark#UpdateScope() and s:MarkScope() by making mark#UpdateMark() take optional (indices, expr) arguments
That were previously directly passed to s:MarkMatch(). This way, there's only one implementation of the iteration in mark#UpdateScope(), and depending on whether arguments are passed in all marks or only the specified indices are updated.
I first tried abstrating the window iteration via a separate s:Windo() function that took a Funcref, but that somehow caused recursive sourcing of the autoload script, because mark#UpdateScope() is invoked at the end of the autoload script. Anyway, the current logic is still a bit too convoluted for my taste, but the separation of functionality actually is quite nice this way, and even backwards compatibility is maintained.