ENH: Add g:mwExclusionPredicates to be able to suppress mark highlighting in window(s)
Now that the mark highlighting update goes through the single mark#UpdateMark() function, we can add an evaluation of user-configured Predicates (as expressions or Funcrefs). If the current window should be skipped for highlighting (for the plugin, the marks are still "there" and can be added / removed, and jumped to, even though there're invisible), existing match highlighting is removed instead of updated.
By default, consider boolean "nomarks" flag variables for various scopes, as these (especially the buffer-local b:nomarks) can be easily set in filetype plugins.
To make this work with changing buffers, we have to run an additional unconditional mark update on BufWinEnter.

Resolves #2.