Compatibility: FIX: Vim versions prior to 7.2.061 choke on g:mwExclusionPredicates
These versions source the autoload script on Funcref definition, and with the default g:mwExclusionPredicates, that loads autoload/mark.vim before the configuration has been fully defined, a Catch-22.
As there are only few and severely outdated 7.1/2 versions affected, and the exclusions are a very recent feature, simply disable the default exclusion. If someone really needs it, after/plugin/mark.vim could be used to define it after the plugin has fully loaded.